delete all customers work now

+++ application/views/customer/index.php
@@ -50,20 +50,22 @@ a:visited:hover{
 	echo "<table>";
 	echo "<tr><th>First</th><th>LAst</th><th>login</th><th>password</th><th>email</th></tr>";
 	
-	foreach ($customers as $customer) {
-		echo "<tr>";
-		echo "<td>" . $customer->first . "</td>";
-		echo "<td>" . $customer->last . "</td>";
-		echo "<td>" . $customer->login . "</td>";
-		echo "<td>" . $customer->password . "</td>";
-		echo "<td>" . $customer->email . "</td>";
+	if (isset($customers)){
+		foreach ($customers as $customer) {
+			echo "<tr>";
+			echo "<td>" . $customer->first . "</td>";
+			echo "<td>" . $customer->last . "</td>";
+			echo "<td>" . $customer->login . "</td>";
+			echo "<td>" . $customer->password . "</td>";
+			echo "<td>" . $customer->email . "</td>";
+			
 		
-	
-		echo "<td>" . anchor("customer/delete/$customer->id",'Delete',"onClick='return confirm(\"Do you really want to delete this record?\");'") . "</td>";
-//		echo "<td>" . anchor("customer/editForm/$customer->id",'Edit') . "</td>";
-		echo "<td>" . anchor("customer/read/$customer->id",'View') . "</td>";
-	
-		echo "</tr>";
+			echo "<td>" . anchor("customer/delete/$customer->id",'Delete',"onClick='return confirm(\"Do you really want to delete this record?\");'") . "</td>";
+	//		echo "<td>" . anchor("customer/editForm/$customer->id",'Edit') . "</td>";
+			echo "<td>" . anchor("customer/read/$customer->id",'View') . "</td>";
+		
+			echo "</tr>";
+		}	
 	}
 	echo "<table>";
 	echo '<p id="icon"><b><span class="logout">' . anchor("login/logout", "Logout") . '</span></p>';
